+++ client/src/scripts/objects/obstacle.ts
@@ -1,9 +1,10 @@
-import { ABParticle2D, Camera2D, ClientParticle2D, ColorM, Container2D, NetStream, NullHitbox2D, ParticlesEmitter2D, random, Sound, Sprite2D, v2 } from "common/engine/client.ts";
-import { Materials, ObstacleDef } from "common/scripts/definitions/objects/obstacles.ts";
+import { ABParticle2D, Camera2D, ClientParticle2D, ColorM, Container2D, NetStream, NullHitbox2D, ParticlesEmitter2D, random, Sound, Sprite2D, type Tween, v2 } from "common/engine/client.ts";
+import { Materials, ObstacleBehaviorDoor, ObstacleDef, ObstacleDoorData } from "common/scripts/definitions/objects/obstacles.ts";
 import { GameObjectType, zIndexes } from "common/scripts/others/constants.ts";
 import { GraphicsDConfig } from "../others/config.ts";
 import { StaticBody, StaticBodyAssetData, StaticBodyPhysicalData } from "./static_body.ts";
 import { Human } from "./human.ts";
+import { CalculateDoorHitbox } from "common/scripts/others/functions.ts";
 export function GetObstacleBaseFrame(def:ObstacleDef,variation:number,skin:number):string{
     let spr=def.assets?.frame?.base??def.idString
 
@@ -24,14 +25,15 @@ export class Obstacle extends StaticBody{
     }={
         hitbox:new NullHitbox2D(v2.zero),
 
-        no_bullets_collision:false,
-        no_collision:false,
+        no_bullets_collision:true,
+        no_collision:true,
         reflect_bullets:false,
 
         scale:0,
         rotation:0,
         side:0
     }
+    door_data?:ObstacleDoorData&{tween?:Tween<any>}
     def!:ObstacleDef
 
     container:Container2D=new Container2D()
@@ -142,9 +144,8 @@ export class Obstacle extends StaticBody{
     }
     set_definition(def:ObstacleDef){
         if(this.def)return
-        if(def.hitbox)this.base_hitbox=def.hitbox.clone()
-
         this.def=def
+
         if(this.def.assets?.sounds){
             this.assets_data.sounds={
                 break:this.game.resources.get_audio(this.def.assets.sounds.break),
@@ -213,13 +214,67 @@ export class Obstacle extends StaticBody{
                 })
             }
         }
-        if(this.def.expanded_behavior&&this.def.hitbox){
+
+        this.physical_data.no_collision=this.def.no_collision??false
+        this.physical_data.no_bullets_collision=this.def.no_bullets_collision??false
+        this.physical_data.reflect_bullets=this.def.reflect_bullets??false
+    }
+    initialize_hitboxes(){
+        if(this.def.hitbox)this.physical_data.hitbox=this.def.hitbox.transform(undefined,undefined,undefined,this.physical_data.side)
+
+        this.base_hitbox=this.physical_data.hitbox
+        if(this.def.expanded_behavior){
             switch(this.def.expanded_behavior.type){
                 case 0:
-                    //this.doors_hitboxes=CalculateDoorHitbox(this.def.hitbox!.to_rect(),this.side,this.def.expanded_behavior as ObstacleBehaviorDoor)
+                    this.door_data={
+                        hitboxes:CalculateDoorHitbox(this.physical_data.hitbox,this.def.expanded_behavior),
+                        locked:false,
+                        open:0,
+                        opening:false,
+                    }
+            }
+        }
+
+        this.manager.cells.updateObject(this)
+    }
+    update_door(ne:number,force:boolean=false){
+        const old=this.door_data!.open
+
+        if(ne!==old){
+            this.door_data!.open=ne as -1|0|1
+
+            if(this.door_data!.tween)this.door_data!.tween.kill()
+
+            let new_rot=this.physical_data.rotation
+
+            if(ne===1)new_rot-=(Math.PI/2)
+            if(ne===-1)new_rot+=(Math.PI/2)
+
+            if(force){
+                this.container.rotation=new_rot
+            }else{
+                if(ne===0&&(this.def.expanded_behavior as ObstacleBehaviorDoor).close_sound){
+                    this.game.sounds.play(this.game.resources.get_audio((this.def.expanded_behavior as ObstacleBehaviorDoor).close_sound!),{
+                        max_distance:30,
+                    },"obstacles")
+                }else if((this.def.expanded_behavior as ObstacleBehaviorDoor).open_sound){
+                    this.game.sounds.play(this.game.resources.get_audio((this.def.expanded_behavior as ObstacleBehaviorDoor).open_sound!),{
+                        max_distance:30,
+                    },"obstacles")
+                }
+
+                this.door_data!.tween=this.game.add_tween({
+                    duration:(this.def.expanded_behavior as ObstacleBehaviorDoor).open_duration,
+                    target:this.container,
+                    to:{rotation:new_rot},
+                    onComplete:()=>{
+                        this.door_data!.tween=undefined
+                    }
+                })
             }
         }
     }
+
     override interact(h:Human){
         if(this.def.expanded_behavior){
             if(this.def.expanded_behavior.type==1){
@@ -251,12 +306,15 @@ export class Obstacle extends StaticBody{
     override auto_interact(h:Human): boolean {
         return (this.def.interactDestroy===true)
     }
+
     override decode(stream: NetStream, full: boolean): void {
         const [
             visual,
             physical_data,physical_data_part,
             health_data,
             dead,
+
+            door_dirty
         ]=stream.readBooleanGroup()
         if(visual||full){
             this.variation=stream.readUint8()
@@ -274,7 +332,8 @@ export class Obstacle extends StaticBody{
                 this.position=stream.readPos2()
                 this.physical_data.rotation=stream.readRad()
                 this.physical_data.side=stream.readUint8()
-                this.manager.cells.updateObject(this)
+
+                this.initialize_hitboxes()
             }
 
             this.container.scale.x=this.physical_data.scale
@@ -295,6 +354,10 @@ export class Obstacle extends StaticBody{
             if(this.emitter_1&&this.health_data.health<=0.4){
                 this.emitter_1.enabled=true
             }
+        }
+
+        if(door_dirty){
+            this.update_door(stream.readInt8(),full)
         }
     }
 }