+++ common/scripts/others/functions.ts
@@ -1,14 +1,11 @@
-import { Orientation, RectHitbox2D, v2 } from "../../engine/core.ts";
+import { Hitbox2D } from "../../engine/core.ts";
 import { ObstacleBehaviorDoor } from "../definitions/objects/obstacles.ts";
 
-export function CalculateDoorHitbox(hitbox:RectHitbox2D,side:Orientation,door:ObstacleBehaviorDoor):Record<-1|0|1,RectHitbox2D>{
+export function CalculateDoorHitbox(hitbox:Hitbox2D,door:ObstacleBehaviorDoor):Record<-1|0|1,Hitbox2D>{
     return {
-        [-1]:hitbox,
+        [-1]:hitbox.transform(undefined,undefined,undefined,1),
         0:hitbox,
-        1:hitbox.transform(
-            v2(3,0),
-            1
-        )
+        1:hitbox.transform(undefined,undefined,undefined,3)
     }
 }
 export function CalculatePlayerLevel(xp:number,base_xp:number=10,factor:number=1.5):number{